src/requirements.txt
gradio>=6.0
langchain>=1.2,<1.3
langchain-openai>=0.3,<1.0
langgraph>=1.0.8,<1.3
langgraph-checkpoint-postgres>=0.0.10
psycopg[binary]>=3.2,<4.0
asyncpg>=0.29,<1.0
python-dotenv
pydantic
pydantic-settings>=2.0
